Fix data type in reference result calculation for consistency in test_tilelang_kernel_gemm_mma_intrinsic.py

+++ testing/python/kernel/test_tilelang_kernel_gemm_mma_intrinsic.py
@@ -222,7 +222,7 @@ def assert_tl_matmul_correctness(M, N, K, in_dtype, out_dtype, accum_dtype):
     assert latency is not None
 
     # Get Reference Result
-    ref_c = torch.matmul(A.to(accum_dtype), B.T.to(accum_dtype)).to(out_dtype)
+    ref_c = torch.matmul(A.to(torch.float32), B.T.to(torch.float32)).to(out_dtype)
     print(C)
     print(ref_c)
     torch.testing.assert_close(C, ref_c, rtol=1e-2, atol=1e-2)
